Добавлены функции верхнего уровня для работы с CurrentBoard

+++ Firmware/Source/Controller/HighLevel/Common.h
@@ -42,8 +42,8 @@ typedef struct __SlaveNode
 
 typedef struct __VIPair
 {
-	float Voltage;
-	float Current;
+	uint32_t Voltage;
+	uint32_t Current;
 
 } VIPair, *pVIPair;
